commit change

+++ Brick.pde
@@ -10,11 +10,11 @@ void display(int colorB){
   fill(255);
   stroke(colorB);
 for(int i =0;i<num;i++){
-  int row =i /NumInRow;
-  int col = i% NumInRow;
+  int row =int(i /NumInRow);
+  int col =int(i% NumInRow);
    
-   int x = IX + brickWidth * col;
-   int y = IY + brickHeight *row;
+   int x = int(IX + brickWidth * col);
+   int y = int(IY + brickHeight *row);
    rect (x,y, brickWidth,brickHeight);
    
   
@@ -32,3 +32,4 @@ num = 50;
 
 
 }
+
